7380 integration test update
