+++ web-client/integration-tests/docketClerkAddsAutomaticBlockCaseToTrial.test.js
@@ -1,0 +1,55 @@
+import { docketClerkCreatesATrialSession } from './journey/docketClerkCreatesATrialSession';
+import { docketClerkSetsCaseReadyForTrial } from './journey/docketClerkSetsCaseReadyForTrial';
+import { docketClerkViewsTrialSessionList } from './journey/docketClerkViewsTrialSessionList';
+import { fakeFile, loginAs, setupTest } from './helpers';
+import { formattedCaseDetail as formattedCaseDetailComputed } from '../src/presenter/computeds/formattedCaseDetail';
+import { petitionsClerkCreatesACaseDeadline } from './journey/petitionsClerkCreatesACaseDeadline';
+import { petitionsClerkCreatesNewCase } from './journey/petitionsClerkCreatesNewCase';
+import { petitionsClerkManuallyAddsCaseToTrial } from './journey/petitionsClerkManuallyAddsCaseToTrial';
+import { runCompute } from 'cerebral/test';
+import { withAppContextDecorator } from '../src/withAppContext';
+
+const formattedCaseDetail = withAppContextDecorator(
+  formattedCaseDetailComputed,
+);
+
+const test = setupTest();
+
+describe('Adds automatic block case to trial', () => {
+  beforeAll(() => {
+    jest.setTimeout(50000);
+  });
+
+  afterAll(() => {
+    test.closeSocket();
+  });
+
+  const trialLocation = `Charleston, West Virginia, ${Date.now()}`;
+  const overrides = {
+    trialLocation,
+  };
+
+  loginAs(test, 'petitionsclerk@example.com');
+  petitionsClerkCreatesNewCase(test, fakeFile, trialLocation);
+
+  loginAs(test, 'docketclerk@example.com');
+  docketClerkSetsCaseReadyForTrial(test);
+  loginAs(test, 'docketclerk@example.com');
+  docketClerkCreatesATrialSession(test, overrides);
+  docketClerkViewsTrialSessionList(test);
+
+  loginAs(test, 'petitionsclerk@example.com');
+  // automatic block with a due date
+  petitionsClerkCreatesACaseDeadline(test);
+  test.casesReadyForTrial = [];
+  petitionsClerkManuallyAddsCaseToTrial(test);
+
+  it('should be able to add a trial session to an automatically blocked case', async () => {
+    const formattedCase = runCompute(formattedCaseDetail, {
+      state: test.getState(),
+    });
+
+    expect(formattedCase.showBlockedTag).toBeTruthy();
+    expect(formattedCase.automaticBlocked).toBeTruthy();
+  });
+});
