+++ web-client/integration-tests/petitionsClerkBlockCase.test.js
@@ -10,7 +10,6 @@ import {
   uploadProposedStipulatedDecision,
   viewCaseDetail,
 } from './helpers';
-import { formattedCaseDetail as formattedCaseDetailComputed } from '../src/presenter/computeds/formattedCaseDetail';
 import { markAllCasesAsQCed } from './journey/markAllCasesAsQCed';
 import { petitionsClerkBlocksCase } from './journey/petitionsClerkBlocksCase';
 import { petitionsClerkCreatesACaseDeadline } from './journey/petitionsClerkCreatesACaseDeadline';
@@ -20,12 +19,6 @@ import { petitionsClerkRemovesPendingItemFromCase } from './journey/petitionsCle
 import { petitionsClerkSetsATrialSessionsSchedule } from './journey/petitionsClerkSetsATrialSessionsSchedule';
 import { petitionsClerkUnblocksCase } from './journey/petitionsClerkUnblocksCase';
 import { petitionsClerkViewsATrialSessionsEligibleCases } from './journey/petitionsClerkViewsATrialSessionsEligibleCases';
-import { runCompute } from 'cerebral/test';
-import { withAppContextDecorator } from '../src/withAppContext';
-
-const formattedCaseDetail = withAppContextDecorator(
-  formattedCaseDetailComputed,
-);
 
 const test = setupTest();
 
@@ -62,16 +55,6 @@ describe('Blocking a Case', () => {
 
   // automatic block with a due date
   petitionsClerkCreatesACaseDeadline(test);
-
-  it('should be able to add a trial session to an automatically blocked case', async () => {
-    const formattedCase = runCompute(formattedCaseDetail, {
-      state: test.getState(),
-    });
-
-    expect(formattedCase.showBlockedTag).toBeTruthy();
-    expect(formattedCase.showNotScheduled).toBeTruthy();
-    expect(formattedCase.automaticBlocked).toBeTruthy();
-  });
 
   it('petitions clerk views blocked report with an automatically blocked case for due date', async () => {
     await refreshElasticsearchIndex();
